Fix clock bug and endpoint bug
